backend/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
openai
openai-whisper
torch
gTTS
